% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_load.R
\name{athena_load}
\alias{athena_load}
\title{Load data to Amazon Athena}
\usage{
athena_load(conn, database, stock)
}
\arguments{
\item{conn}{odbc conn created with mobstr::athena_connect}

\item{database}{string representing the database name in AWS Athena}

\item{stock}{string representing the stock symbol to be uploaded to athena}
}
\description{
{
Make an ODBC connection to an Amazon Athena using settings from
the environment variables loaded with .RProfile
}
}
