% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_leg.R
\name{close_leg}
\alias{close_leg}
\title{Close single leg options trade}
\usage{
close_leg(df, entry_date, exp, typ, stk, entry_mid, entry_delta, stk_price,
  direction)
}
\arguments{
\item{df}{dataframe of opened options legs}

\item{entry_date}{date position was opened}

\item{exp}{date of expiration for position used to filter possible results}

\item{typ}{is this a put or a call option}

\item{stk}{option strike price}

\item{entry_mid}{mid price of the position when opened}

\item{entry_delta}{delta of the strike when position was opened}

\item{stk_price}{stock price when position was opened}

\item{direction}{was this a short or long position used to calc profit}
}
\value{
A dataframe of filtered closing option trades to be used in further
analysis. One row for each day a posistion was open open_date +1 through
expiration
}
\description{
{
Using function inputs from opened trades close each day until expiration
}
}
