% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent_trades.R
\name{concurrent_trades}
\alias{concurrent_trades}
\title{Calculate # concurrent trades for each day}
\usage{
concurrent_trades(sy, st, en)
}
\arguments{
\item{sy}{character string representing stock symbol for group and join later}

\item{st}{character string understood by lubridate ymd to represent open date of trade}

\item{en}{character string understood by lubridate ymd to represent closing date of trade}
}
\value{
dataframe of dates when position was open including stock symbol

head(concurrent_trades("SPY", "2018-01-01", "2018-02-16"))
symbol quote_date  exit_date
SPY 2018-01-01 2018-02-16
SPY 2018-01-02 2018-02-16
SPY 2018-01-03 2018-02-16
SPY 2018-01-04 2018-02-16
SPY 2018-01-05 2018-02-16
SPY 2018-01-06 2018-02-16
}
\description{
{
Number of trades open on any given day can be used to calculate when multiple
entries would cause more capital allocated to a strategy by symbol than desired
}
}
\examples{
concurrent_trades("SPY", "2018-01-01", "2018-02-16")

Function ----
}
