% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lags}
\alias{lags}
\title{Short utility functions for the mobstr package}
\usage{
lags(var, n = 10)
}
\arguments{
\item{var}{variable name to create lags of}

\item{n}{number of lags to create}
}
\value{
A vector to include in a call to mutate in dplyr chain
}
\description{
{
Various utility functions for use with the mobstr package
}
}
