% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_options.R
\name{load_options}
\alias{load_options}
\title{Load options data from local source}
\usage{
load_options(stock, tar_dte)
}
\arguments{
\item{stock}{string representing the stock symbol}

\item{tar_dte}{int representing the DTE (Days to Expiration)}
}
\description{
{
Load options data from local source
}
}
