% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_redshift.R
\name{load_redshift}
\alias{load_redshift}
\title{Load options data into AWS redshift}
\usage{
load_redshift(stock)
}
\arguments{
\item{stock}{string representing the stock symbol}
}
\description{
{
Load options data into redshift table
}
}
