% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_leg.R
\name{open_leg}
\alias{open_leg}
\title{TODO resolve method to use max strike for shorts and min strike for calls
Open single leg options trade}
\usage{
open_leg(conn, stock, put_call, direction, tar_delta, tar_dte)
}
\arguments{
\item{conn}{AWS Athena connection}

\item{stock}{character string of stock symbol}

\item{put_call}{is this a put or call option we are opening?}

\item{direction}{credit or debit (short or long trade)}

\item{tar_delta}{target delta for the option}

\item{tar_dte}{target number of days to expiration at entry}
}
\value{
A dataframe of filtered opening option trades to be used in further
analysis.
}
\description{
{
Using function inputs for options type, direction, target delta, and buffer
open a single option leg trade
}
}
