% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_short_vert.R
\name{open_short_vert}
\alias{open_short_vert}
\title{Open short vertical spread}
\usage{
open_short_vert(data, tar_delta, put_call, width)
}
\arguments{
\item{data}{dataframe of stock options}

\item{tar_delta}{integer of target delta for opening trade}

\item{put_call}{string either put or call}

\item{width}{numeric representing the width of the spread}
}
\description{
{
Open Short Vertical Spread trade
}
}
