% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_connect.R
\name{redshift_connect}
\alias{redshift_connect}
\title{Connect to Amazon Redshift}
\usage{
redshift_connect(env)
}
\arguments{
\item{env}{variable used to create paths for various systems on AWS}
}
\description{
{
Make an RJDBC connection to an Amazon Redshift cluster using settings from
the environment variables loaded with .RProfile
}
}
\details{
Download Amazon Redshift JDBC driver
download.file('http://s3.amazonaws.com/redshift-downloads/drivers/RedshiftJDBC41-1.1.9.1009.jar','RedshiftJDBC41-1.1.9.1009.jar')

Install commandline tools for MAC after updating the OS
https://developer.apple.com/download/more/
}
\examples{
redshift_connect("TASTYTRADE")


}
