% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_redshift.R
\name{truncate_redshift}
\alias{truncate_redshift}
\title{Truncate redshift table}
\usage{
truncate_redshift(connection, table_name)
}
\arguments{
\item{connection}{string representing RJDBC connection name}

\item{table_name}{string representing the table name in redshift}
}
\description{
{
Truncate redshift table
}
}
