% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_leg.R
\name{open_leg}
\alias{open_leg}
\title{Open single leg options trade}
\usage{
open_leg(data, put_call, direction, tar_delta, tar_buffer)
}
\arguments{
\item{data}{table from redshift not yet collected}

\item{put_call}{is this a put or call option we are opening?}

\item{direction}{credit or debit (short or long trade)}

\item{tar_delta}{target delta for the option}

\item{tar_buffer}{percent the strike should be away from ATM}
}
\value{
A dataframe of filtered opening option trades to be used in further
analysis.
}
\description{
{
Using function inputs for options type, direction, target delta, and buffer
open a single option leg trade
}
}
\examples{
open_leg(data = options, put_call = "put", direction = "credit",
         tar_delta = -0.16, tar_buffer = 0.10)


}
