% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_2_s3.R
\name{write_2_S3}
\alias{write_2_S3}
\title{Write to AWS S3}
\usage{
write_2_S3(file, path, bucket)
}
\arguments{
\item{file}{character string of the file we want to put to S3}

\item{path}{character string of local file path to upload}

\item{bucket}{character string of the name for the root AWS S3 bucket}
}
\value{
TRUE when successful copy is made
}
\description{
{
Take processed csv file options file put to AWS S3 bucket
}
}
