% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_2_s3.R
\name{write_2_S3}
\alias{write_2_S3}
\title{Write to AWS S3}
\usage{
write_2_S3(stock)
}
\arguments{
\item{stock}{character string of the stock we want to put to S3}
}
\value{
TRUE when successful copy is made
}
\description{
{
Take processed RDS stock options file and convert to CSV then put to AWS S3
bucket
}
}
